requirements.txt
Django==4.1.5
django-crontab==0.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.5
PyJWT==2.6.0
redis==4.4.1
gunicorn==20.1.0
django-extensions==3.2.1
